fix(diplomacy): split power-score helpers to keep relation-lookup under size budget (Refs #3753)
diplomacy_relation_lookup.dart exceeded the 500-line domain-package source
file budget (repo.domain_package_source_file_size) after the R7.1 favoured
trading partner addition. Extract the Great Power power-score concern into
diplomacy_power_score.dart, re-export via the package barrel, and point the
direct power-score test import at the new file. No behavior change.
