+++ packages/colonizethis_diplomacy/test/diplomacy/calendar_power_score_winner_test.dart
@@ -1,4 +1,4 @@
-import 'package:colonizethis_diplomacy/src/diplomacy/diplomacy_relation_lookup.dart';
+import 'package:colonizethis_diplomacy/src/diplomacy/diplomacy_power_score.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
